Moved Navbar and footer to their appropriate places, remake the scrolltotop component
