+++ frontend/src/components/styles/Contributors.css
