+++ frontend/src/components/styles/shared-layout.css
